+++ src/main/java/com/jin/tpdb/persistence/DAO.java
@@ -86,7 +86,7 @@ public class DAO {
 	}
 	
 	
-    public int getAlbumTotalComments(int id) {
+    public Long getAlbumTotalComments(int id) {
 
 		/*$albums_query = "SELECT users.username, albums.album_id, albums.uploader_user_id,	albums.album_name,
 						albums.upload_date, albums.cover, albums.description, artists.artist_name,
@@ -96,7 +96,7 @@ public class DAO {
 		*/
 		String qs = "SELECT COUNT(ac) FROM AlbumComment ac";
 		TypedQuery<Long> query = em.createQuery(qs, Long.class);
-		return (int)query.getSingleResult();		
+		return query.getSingleResult();		
 		/*Root<AlbumComment> root = cq.from(AlbumComment.class);		
 		cq.select(qb.count(root));
 		Predicate predicate = qb.equal(root.get("album_id"), id);
